# backend/docs/swagger.yaml
basePath: /
definitions:
  protocol.GameRoomStats:
    properties:
      bestOverallDuration:
        type: integer
      bestTodaysDuration:
        type: integer
      countActive:
        type: integer
      countLeaderboard:
        type: integer
    type: object
host: buttonmania.win
info:
  contact:
    email: team@buttonmania.win
    name: ButtonMania Team
  title: ButtonMania API
  version: "1.0"
paths:
  /api/room/create:
    get:
      parameters:
      - description: Client ID
        in: query
        name: clientId
        required: true
        type: string
      - description: Room ID
        in: query
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
        "400":
          description: Room exists
      summary: Create game room
  /api/room/delete:
    get:
      parameters:
      - description: Client ID
        in: query
        name: clientId
        required: true
        type: string
      - description: Room ID
        in: query
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
        "400":
          description: Room cannot be deleted
        "404":
          description: Room not found
      summary: Delete game room
  /api/room/stats:
    get:
      parameters:
      - description: Client ID
        in: query
        name: clientId
        required: true
        type: string
      - description: Room ID
        in: query
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/protocol.GameRoomStats'
        "400":
          description: Room id is too long
        "404":
          description: Room not found
      summary: Get room stats
  /ws:
    get:
      parameters:
      - description: Client ID
        in: query
        name: clientId
        required: true
        type: string
      - description: Room ID
        in: query
        name: roomId
        required: true
        type: string
      - description: Telegram init data
        in: query
        name: initData
        required: true
        type: string
      responses: {}
      summary: Handles WebSocket connections
swagger: "2.0"